refactor(cmd): remove import aliases
