+++ cmd/config.go
@@ -3,10 +3,14 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 
+	"github.com/alex-held/devctl/internal/devctlpath"
+
+	"github.com/alex-held/devctl/internal/config"
+
 	"github.com/alex-held/devctl/internal/cli"
-	config2 "github.com/alex-held/devctl/internal/config/config"
 )
 
 func NewConfigCommand() *cobra.Command {
@@ -42,8 +46,10 @@ This application is a tool to generate the needed files
 to quickly create a Cobra application.`,
 		Args: cobra.ExactArgs(0),
 		Run: func(cmd *cobra.Command, args []string) {
-			devEnvConfig := config2.LoadViperConfig()
-			configString := fmt.Sprintf("%+v\n", *devEnvConfig)
+			devEnvConfig := config.LoadViperConfig()
+			cfg, err := config.Load(afero.NewOsFs(), devctlpath.NewPather())
+			cli.ExitWithError(1, err)
+			configString := fmt.Sprintf("NewConfig: %+v\nOldConfig: %+v\n", *cfg, *devEnvConfig)
 			fmt.Println(configString)
 		},
 	}
